# Add protocol buffers for binary injection/joining