--- a/wire/wire.proto
+++ b/wire/wire.proto
@@ -1,0 +1,10 @@
+syntax = "proto3";
+package basictracer_go.wire;
+option go_package = "wire";
+
+message TracerState {
+  sfixed64 trace_id = 1;
+  sfixed64 span_id = 2;
+  bool    sampled = 3;
+  map<string, string> baggage_items = 4;
+}