* Replace the internal use of `encoding/binary` inside of `(*splitBinaryPropagator).Join` and `(*splitBinaryPropagator).Inject` with Google's Protocol buffers. Gogoproto is used for the go code generation since it provides considerable performance gains over vanilla protobufs.

* Adds a ProtobufCarrier to allow end users to use other protobuf encoding formats, i.e. jsonpb.

The benchmarks were run on an AWS m3.large. 
```
name                                           old time/op    new time/op    delta
Span_Empty                                        710ns ±55%     657ns ± 0%   -7.53%          (p=0.037 n=8+7)
Span_100Events                                   48.6µs ± 1%    49.9µs ± 0%   +2.66%         (p=0.000 n=9+10)
Span_1000Events                                  48.5µs ± 0%    49.9µs ± 0%   +2.85%        (p=0.000 n=10+10)
Span_100Tags                                     47.0µs ± 1%    46.5µs ± 2%   -0.91%        (p=0.023 n=10+10)
Span_1000Tags                                    46.8µs ± 1%    47.0µs ± 1%     ~           (p=0.529 n=10+10)
Span_100BaggageItems                              155µs ± 0%     157µs ± 0%   +1.30%        (p=0.000 n=10+10)
TrimmedSpan_100Events_100Tags_100BaggageItems     181µs ± 0%     181µs ± 0%   +0.36%        (p=0.000 n=10+10)
Inject_SplitText_Empty                            468ns ± 1%     476ns ± 1%   +1.56%        (p=0.000 n=10+10)
Inject_SplitText_100BaggageItems                 20.0µs ± 2%    20.1µs ± 1%     ~           (p=0.089 n=10+10)
Inject_GoHTTPHeader_Empty                        2.98µs ± 4%    1.39µs ± 1%  -53.33%        (p=0.000 n=10+10)
Inject_GoHTTPHeader_100BaggageItems              70.7µs ± 1%    48.4µs ± 1%  -31.57%          (p=0.000 n=8+9)
Inject_SplitBinary_Empty                         1.30µs ± 2%    0.31µs ± 1%  -76.08%         (p=0.000 n=10+9)
Inject_SplitBinary_100BaggageItems               52.8µs ± 1%    36.0µs ± 1%  -31.94%        (p=0.000 n=10+10)
Join_SplitText_Empty                             1.61µs ± 0%    1.62µs ± 1%   +0.86%        (p=0.000 n=10+10)
Join_SplitText_100BaggageItems                   54.0µs ± 2%    53.8µs ± 1%     ~           (p=0.739 n=10+10)
Join_GoHTTPHeader_Empty                          2.52µs ± 1%    1.62µs ± 0%  -35.97%         (p=0.000 n=10+8)
Join_GoHTTPHeader_100BaggageItems                 152µs ± 2%      91µs ± 2%  -40.10%        (p=0.000 n=10+10)
Join_SplitBinary_Empty                           1.73µs ± 1%    0.96µs ± 1%  -44.35%        (p=0.000 n=10+10)
Join_SplitBinary_100BaggageItems                 96.8µs ± 1%    42.4µs ± 4%  -56.15%         (p=0.000 n=9+10)

name                                           old alloc/op   new alloc/op   delta
Span_Empty                                       0.75B ±700%    0.00B ±NaN%     ~             (p=1.059 n=8+9)
Span_100Events                                   16.1kB ± 0%    16.1kB ± 0%     ~     (all samples are equal)
Span_1000Events                                  16.1kB ± 0%    16.1kB ± 0%     ~     (all samples are equal)
Span_100Tags                                     10.5kB ± 0%    10.5kB ± 0%     ~           (p=0.355 n=10+10)
Span_1000Tags                                    10.5kB ± 0%    10.5kB ± 0%     ~           (p=0.692 n=10+10)
Span_100BaggageItems                             10.5kB ± 0%    10.5kB ± 0%     ~           (p=0.724 n=10+10)
TrimmedSpan_100Events_100Tags_100BaggageItems    0.00B ±NaN%    0.00B ±NaN%     ~     (all samples are equal)
Inject_SplitText_Empty                            48.0B ± 0%     48.0B ± 0%     ~     (all samples are equal)
Inject_SplitText_100BaggageItems                 3.25kB ± 0%    3.25kB ± 0%     ~     (all samples are equal)
Inject_GoHTTPHeader_Empty                          641B ± 0%      321B ± 0%  -49.92%        (p=0.000 n=10+10)
Inject_GoHTTPHeader_100BaggageItems              30.1kB ± 0%    18.9kB ± 0%  -37.18%         (p=0.000 n=9+10)
Inject_SplitBinary_Empty                           368B ± 0%       80B ± 0%  -78.26%        (p=0.000 n=10+10)
Inject_SplitBinary_100BaggageItems               6.77kB ± 0%    6.39kB ± 0%   -5.56%        (p=0.000 n=10+10)
Join_SplitText_Empty                              16.0B ± 0%     16.0B ± 0%     ~     (all samples are equal)
Join_SplitText_100BaggageItems                   10.5kB ± 0%    10.5kB ± 0%     ~            (p=0.823 n=10+9)
Join_GoHTTPHeader_Empty                            288B ± 0%       80B ± 0%  -72.22%        (p=0.000 n=10+10)
Join_GoHTTPHeader_100BaggageItems                33.3kB ± 0%    27.8kB ± 0%  -16.57%        (p=0.000 n=10+10)
Join_SplitBinary_Empty                             240B ± 0%       48B ± 0%  -80.00%        (p=0.000 n=10+10)
Join_SplitBinary_100BaggageItems                 22.5kB ± 0%    17.0kB ± 0%  -24.52%        (p=0.000 n=10+10)

name                                           old allocs/op  new allocs/op  delta
Span_Empty                                        0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
Span_100Events                                     8.00 ± 0%      8.00 ± 0%     ~     (all samples are equal)
Span_1000Events                                    8.00 ± 0%      8.00 ± 0%     ~     (all samples are equal)
Span_100Tags                                       11.0 ± 0%      11.0 ± 0%     ~     (all samples are equal)
Span_1000Tags                                      11.0 ± 0%      11.0 ± 0%     ~     (all samples are equal)
Span_100BaggageItems                               11.0 ± 0%      11.0 ± 0%     ~     (all samples are equal)
TrimmedSpan_100Events_100Tags_100BaggageItems     0.00 ±NaN%     0.00 ±NaN%     ~     (all samples are equal)
Inject_SplitText_Empty                             3.00 ± 0%      3.00 ± 0%     ~     (all samples are equal)
Inject_SplitText_100BaggageItems                    103 ± 0%       103 ± 0%     ~     (all samples are equal)
  ject_GoHTTPHeader_Empty                          15.0 ± 0%       5.0 ± 0%  -66.67%        (p=0.000 n=10+10)
Inject_GoHTTPHeader_100BaggageItems                 421 ± 0%        11 ± 0%  -97.39%        (p=0.000 n=10+10)
Inject_SplitBinary_Empty                           11.0 ± 0%       3.0 ± 0%  -72.73%        (p=0.000 n=10+10)
Inject_SplitBinary_100BaggageItems                  411 ± 0%         7 ± 0%  -98.30%        (p=0.000 n=10+10)
Join_SplitText_Empty                               1.00 ± 0%      1.00 ± 0%     ~     (all samples are equal)
Join_SplitText_100BaggageItems                     12.0 ± 0%      12.0 ± 0%     ~     (all samples are equal)
Join_GoHTTPHeader_Empty                            13.0 ± 0%       3.0 ± 0%  -76.92%        (p=0.000 n=10+10)
Join_GoHTTPHeader_100BaggageItems                   622 ± 0%       217 ± 0%  -65.11%        (p=0.000 n=10+10)
Join_SplitBinary_Empty                             11.0 ± 0%       2.0 ± 0%  -81.82%        (p=0.000 n=10+10)
Join_SplitBinary_100BaggageItems                    619 ± 0%       214 ± 0%  -65.43%        (p=0.000 n=10+10)
```