@bensigelman 2016-03-08
```
Inject_GoHTTPHeader_100BaggageItems                 421 ± 0%        11 ± 0%  -97.39%        (p=0.000 n=10+10)
```

^^^ nice!!