@bg451 2016-03-08
On a side note, does having the protobuf code in the package `wire` make sense? Something about it just feels off to me, but I'm not sure why.