@bensigelman 2016-03-08
> On a side note, does having the protobuf code in the package `wire` make sense? Something about it just feels off to me, but I'm not sure why.

Per that comment I made, I think the .proto belongs somewhere far away (probably in a different repo altogether)... I can see lots of alternatives to `wire` but I was having trouble coming up with strong arguments one way or the other.