@bg451 2016-03-08
I agree that it should be moved to something like `basictracer-common`. I suppose for now we can leave it in here until OT 1.0 becomes official or someone starts working on another basictracer implementation.