@bg451 2016-03-22
Rebased and squashed everything now that #15 is in.

Here's the updated performance numbers:
```
name                                             old time/op    new time/op    delta
Inject_Binary_Empty-2                              1.03µs ± 0%    0.49µs ± 0%  -51.95%         (p=0.000 n=10+7)
Inject_Binary_100BaggageItems-2                    61.0µs ± 1%    16.0µs ± 2%  -73.85%         (p=0.000 n=10+9)
Join_Binary_Empty-2                                1.75µs ± 1%    1.49µs ± 0%  -15.14%        (p=0.000 n=10+10)
Join_Binary_100BaggageItems-2                      95.4µs ± 0%    46.6µs ± 1%  -51.18%        (p=0.000 n=10+10)

name                                             old alloc/op   new alloc/op   delta
Inject_Binary_Empty-2                                219B ± 0%      168B ± 0%  -23.29%        (p=0.000 n=10+10)
Inject_Binary_100BaggageItems-2                    23.8kB ± 1%    16.3kB ± 0%  -31.34%        (p=0.000 n=10+10)
Join_Binary_Empty-2                                  256B ± 0%      224B ± 0%  -12.50%        (p=0.000 n=10+10)
Join_Binary_100BaggageItems-2                      22.5kB ± 0%    21.7kB ± 0%   -3.49%         (p=0.000 n=10+9)

name                                             old allocs/op  new allocs/op  delta
Inject_Binary_Empty-2                                9.00 ± 0%      5.00 ± 0%  -44.44%        (p=0.000 n=10+10)
Inject_Binary_100BaggageItems-2                       409 ± 0%         5 ± 0%  -98.78%        (p=0.000 n=10+10)
Join_Binary_Empty-2                                  10.0 ± 0%       6.0 ± 0%  -40.00%        (p=0.000 n=10+10)
Join_Binary_100BaggageItems-2                         618 ± 0%       217 ± 0%  -64.89%        (p=0.000 n=10+10)