@bg451 2016-03-22
cc @bensigelman @tschottdorf 

Is there any way to get `golint` to ignore specific files? It's not happy with the generated protobuf code.