@bensigelman 2016-03-22
@bg451 basically LGTM with a few caveats:
- we will need to move the proto to `basictracer-common` when we have a second `basictracer` platform
- there's the silly mechanical issue of getting TravisCI to ignore the genfiles
- I am wondering about how to represent the "sampled" state in a forwards-compatible manner...