@bensigelman 2016-03-24
This LGTM, though it would be nice if @tschottdorf took a look since he is using `basictracer-go` at cockroach.