@tschottdorf 2016-03-24
LGTM, but needs a rebase and squash.