@bg451 2016-03-24
PTAL. Comments addressed, squashed and rebased.