@bensigelman 2016-03-24
LGTM though I think we should give @tschottdorf a chance to take a look since he had mentioned the varint thing.