@tschottdorf 2016-03-24
All good, nothing I'm concerned about.

On Thu, Mar 24, 2016, 19:52 bhs <notifications@github.com> wrote:

> LGTM though I think we should give @tschottdorf
> <https://github.com/tschottdorf> a chance to take a look since he had
> mentioned the varint thing.
>
> —
> You are receiving this because you were mentioned.
>
> Reply to this email directly or view it on GitHub
> <https://github.com/opentracing/basictracer-go/pull/13#issuecomment-201076582>
>